perf(test): split the six build-running artifact tests into parallel files
tests/artifacts.test.ts was 118.8s of the 144.3s shared-registry pass -- 82% of
CI's test floor in one file. Six of its tests each ran a full
scripts/build-artifacts.ts (7.2s-24.7s apiece, 99.4s combined), and vitest
parallelizes across FILES, never within one; test.concurrent could not help
either, because execFileSync blocks the whole worker. The only arrangement
that overlaps the builds is separate files on separate workers.

Move the four heaviest into sibling files -- artifacts-build-health,
artifacts-build-schema, artifacts-build-determinism, artifacts-build-schema-
index -- each with its own repo sandbox via the new
tests/helpers/artifact-build-harness.ts factory (per-file sandbox + path
helpers + support-artifact snapshot/restore, extracted from artifacts.test.ts).
The extra ~0.6s clone per file is noise against a ~20s build. Locally the five
files now finish in 61.5s wall with 165.8s of test time inside -- the slowest
file is 38.6s, down from 118.8s.

The split surfaced a hidden in-file ordering dependency: the llms.txt test
runs the real build, which rm's + repopulates the staging dir WITHOUT an
r2-manifest.json (that file is r2-manifest.ts's product, not the build's), and
the two R2 upload tests 2000 lines later only found a manifest because the
departed build tests' cleanup happened to recreate one. They now regenerate it
themselves (memoized ensureStagingManifest), so they pass in any order and
under vitest -t isolation, which they never did before.

The moved tests' explicit timeouts go from 30s to 120s: they measured
21-25s in CI against the old 30s cap, a margin that already blew up on a
loaded machine.
